Splitting up the SlideTransformer tests into multiple files
